Generate repository.xml
The release script expects two things from sdk-repo-$OS-tools-$NUM.zip:
- they should contain a source.properties
- everything should be inside a tools folder.
- finally a repository.xml should be created, based on the source.properties.

Eventually we should do all of this from within gradle.
Since I'm not familiar with that part of the code,
as a first step, fix the zip files directly in the build_tools.sh
script and invoke the same mk_sdk_repo_xml.sh as present in
development.git.

Change-Id: Iab90b35e93daf0a1272b3a9d102cbd10d515eeff