Adding debug file and zip target to sdk tools
Tool items now have a debug attribute that will place them in a
tools.debug root directory. The tools.debug directory will be zipped
separately and copied to the distribution directory for storage.

Also added an accessor to private itemPath so it would show up in
ToolItem's property map

Emulator uses this to store debug symbols that match the binaries
distributed in tools that will not be released.

Change-Id: I5cbc52554b8f91424a92b525211632c213dbf174