Run unit tests simultaneously on build bots
We have two sets of build machines:

 1. Machines controlled by ourselves.
 2. Machines controlled by the build team. (Darwin builds)

For 1 we have the guarantee that no more than one build is running
concurrently, and we are not concerned about cpu load. Case 2 is an
unknown.

We have seen test failures due to heavy cpu load (mostly timeouts).
Since we know the cpu load in case 1, we are willing to run the tests
concurrently. For case 2 we don't so we disable it.

Once we have our own fleet of Darwin machines we can start increasing
the job counts for tests as well.

Bug: b/128613167
Change-Id: I027afb05cf3159f8b931a01a82bde7a2ef032d5e