Improve python setup times and fix site dir creation.
Python does a files system scan during the exection of setup.py. This
scan will go through cwd until it find the packages of interest.

We now run setup.py from the directory where it is located so the scan
will contain exactly what is needed.

Fresh windows build bots do not have a site-packages directory
initialized, and it seems that setuptools will not properly create it,
causing trouble.

Change-Id: I76e972a71f88377b01535b91aaf6d0cd56f9f5cf